// FrontEnd/packages/api/generated/api.ts
export class LoginRequestDto implements ILoginRequestDto {
    email?: string | undefined;
    password?: string | undefined;

    constructor(data?: ILoginRequestDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.email = _data["email"];
            this.password = _data["password"];
        }
    }

    static fromJS(data: any): LoginRequestDto {
        data = typeof data === 'object' ? data : {};
        let result = new LoginRequestDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["email"] = this.email;
        data["password"] = this.password;
        return data;
    }
}

export interface ILoginRequestDto {
    email?: string | undefined;
    password?: string | undefined;
}
// ...
export class LoginTotpRequestDto implements ILoginTotpRequestDto {
    email?: string | undefined;
    password?: string | undefined;
    code?: string | undefined;

    constructor(data?: ILoginTotpRequestDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.email = _data["email"];
            this.password = _data["password"];
            this.code = _data["code"];
        }
    }

    static fromJS(data: any): LoginTotpRequestDto {
        data = typeof data === 'object' ? data : {};
        let result = new LoginTotpRequestDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["email"] = this.email;
        data["password"] = this.password;
        data["code"] = this.code;
        return data;
    }
}

export interface ILoginTotpRequestDto {
    email?: string | undefined;
    password?: string | undefined;
    code?: string | undefined;
}
// ...
export class RefreshRequestDto implements IRefreshRequestDto {
    refreshToken?: string | undefined;

    constructor(data?: IRefreshRequestDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.refreshToken = _data["refreshToken"];
        }
    }

    static fromJS(data: any): RefreshRequestDto {
        data = typeof data === 'object' ? data : {};
        let result = new RefreshRequestDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["refreshToken"] = this.refreshToken;
        return data;
    }
}

export interface IRefreshRequestDto {
    refreshToken?: string | undefined;
}
// ...
export class TokenResponseDto implements ITokenResponseDto {
    accessToken?: string | undefined;
    accessTokenExpiresAtUtc?: Date;
    refreshToken?: string | undefined;
    refreshTokenExpiresAtUtc?: Date;

    constructor(data?: ITokenResponseDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.accessToken = _data["accessToken"];
            this.accessTokenExpiresAtUtc = _data["accessTokenExpiresAtUtc"] ? new Date(_data["accessTokenExpiresAtUtc"].toString()) : undefined as any;
            this.refreshToken = _data["refreshToken"];
            this.refreshTokenExpiresAtUtc = _data["refreshTokenExpiresAtUtc"] ? new Date(_data["refreshTokenExpiresAtUtc"].toString()) : undefined as any;
        }
    }

    static fromJS(data: any): TokenResponseDto {
        data = typeof data === 'object' ? data : {};
        let result = new TokenResponseDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["accessToken"] = this.accessToken;
        data["accessTokenExpiresAtUtc"] = this.accessTokenExpiresAtUtc ? this.accessTokenExpiresAtUtc.toISOString() : undefined as any;
        data["refreshToken"] = this.refreshToken;
        data["refreshTokenExpiresAtUtc"] = this.refreshTokenExpiresAtUtc ? this.refreshTokenExpiresAtUtc.toISOString() : undefined as any;
        return data;
    }
}

export interface ITokenResponseDto {
    accessToken?: string | undefined;
    accessTokenExpiresAtUtc?: Date;
    refreshToken?: string | undefined;
    refreshTokenExpiresAtUtc?: Date;
}